## Summary

`DEBUG_ASSERTCRASH` expands to `((void)0)` unless `DEBUG_CRASHING` is
defined, and `Debug.h` derives that from `RTS_DEBUG` — which did not
compile off Windows. So **no native figure this project has ever
published came from a build in which one of the engine's assertions
could fire**, which is why every serious defect this wave was silent
(`headless-simulation-probe.md` §3-§4: a retail map parsed, reported
loaded, simulated 13,500 frames with zero objects and a constant CRC,
with the guard for that condition two lines away and compiled out).

Measured first, on this tree, before any fix (`--config debug` =
`-DRTS_DEBUG -DWWDEBUG -DDEBUG`, the defines `cmake/config-build.cmake`
sets for `RTS_BUILD_OPTION_DEBUG=ON`):

```
objects 967/978, compile failures 11, undefined symbols 114, binary produced: no
  compile-blocked 113, no-definition-anywhere 1 (GlobalSize),
  library-not-linked 0, cut-scope-not-linked 0, harness-artefact 0
```

The slice's scoping figure (10 TUs / 113 symbols) predates
#96/#98/#99/#100; the per-TU first diagnostics and the full
categorisation are in `docs/porting/native-debug-build.md` §1. After:
**978/978, 0 failures, 0 unresolved, executable produced**, with release
re-measured unmoved at 978/978 / 0.

Seams, all under the existing spelling so no consumer call site changed.
The non-obvious one is not a shim at all:

```diff
-typedef long long Int64;          // MSVC only
+#ifdef _MSC_VER
+typedef int64_t   Int64;
+#else
+typedef long long Int64;          // the same type __int64 maps to, so the debug
+#endif                           // callers' `__int64*` and `Int64*` are one type at LP64
```

which is what the four `GetPrecisionTimer{,TicksPerSec}` failures
actually were. The rest: `<windows.h>` in `CRCDebug.cpp`
(`CreateDirectory`/`DeleteFile`), POSIX
`GetUserNameA`/`GetComputerNameA`/`MAX_COMPUTERNAME_LENGTH`, `DWORD` not
`unsigned long` for `Recorder.cpp`'s buffer length, portable WinSock
result values for a debug-only log switch (GameSpy stays cut — its
*logging* compiles, nothing else), `Is_Trying_To_Exit` declared off
Windows too, `GlobalSize` over `malloc_size`/`malloc_usable_size` beside
the existing `GlobalAlloc`, and one debug-log statement that indexed a
`const Char*` as a struct because nothing had ever compiled it.

**What an assertion does off Windows now.** The Windows path is an
Abort/Retry/Ignore dialog and the portable stub used to answer it, which
would have made a compiled-in assertion indistinguishable from a
compiled-out one. `DEBUG_CRASH_AT()` records
`__FILE__`/`__LINE__`/`#cond` in `TheCurrentCrash*` next to the existing
`TheCurrentIgnoreCrashPtr`, and off Windows `DebugCrash()` writes to
stderr whether or not the log was initialised, dumps the stack through
the existing `StackDump.cpp` seam (now demangling with
`abi::__cxa_demangle`; Unix debug links `-rdynamic` so frames have
names), and exits non-zero unless `-ignoreAsserts` is passed. The
file/line/condition prefix is `#ifndef _WIN32`: on Windows the dialog is
the report and its text is behaviour the Windows build owns.

**Negative control**, in CI, no retail data needed —
`scripts/ci/check-assert-fires.py` generates a map whose chunk TOC does
not name the ID the file then uses and requires exit≠0, `ASSERTION
FAILURE`, the engine's own file and message, a stack dump, and at least
one *named* frame:

```
ASSERTION FAILURE: .../System/DataChunk.cpp:477: name not found in DataChunkTableOfContents::getName for id 7
Stack Dump:
  .../sim_probe(DataChunkTableOfContents::getName(unsigned int)+0xd0)
  .../sim_probe(DataChunkInput::openDataChunk(unsigned short*)+0x16d)
[Abort: assertion failed off Windows, where there is no assert dialog to ignore it. …]
exit status: 1
```

**Re-running the harness under the debug build against retail data**
(`native-debug-build.md` §5) turned up two things, stated at full
strength:

1. **A port defect an earlier slice introduced, fixed here because
nothing else could run until it was.** `MemoryPool::createBlob` asserts
the blob's block *count* is a multiple of `MEM_BOUND_ALIGNMENT`, which
rose 4 → 16 at 64-bit for a real reason (the allocator must meet
`alignof(max_align_t)`), silently tightening an unrelated *count* check
that the retail pool table violates (`{ "NameKeyBucketPool", 9000, 1024
}` — legal at 4). Every debug run aborted in the first
`NameKeyGenerator` lookup. A count is not an alignment: blocks are
`m_allocationSize` bytes and that is already rounded to the bound. The
check now uses its own `MEM_POOL_COUNT_GRANULARITY` of 4 — the value the
retail tables were authored against and the value `GameMemoryInit.cpp`
still rounds counts to — so the 32-bit Windows predicate is unchanged
and the assertion still rejects zero, negative and off-granularity
counts.
2. **A finding for the filesystem seam, not fixed here, and it is one
slice rather than three items.** `MapUtil.cpp:536` requires a backslash
in an enumerated map path and off Windows enumeration returns forward
slashes: an assertion about a Windows-only assumption, and blocker 5
already. What changed is that it is no longer silent — the same path
previously produced an empty map cache and reported success. Scope it
with the two defects sharing its code path and cause: the `MapCache`
path defect the startup probe reports (`startability.md` §6 item 2 /
`review-and-decisions.md` §2.6) and the `GameSlot::setState` dereference
below.

`chunks`/`filecrc`/`xfercrc` over the retail map fire nothing and return
the release configuration's CRCs (`6314F5FD`, `2465AFE3`); `replayhdr`
still SIGSEGVs in `GameSlot::setState` via `RecorderClass::init` because
the harness never creates `TheGameText` — harness limitation (blocker
10), and notably the debug build adds *nothing* there, since a raw null
deref has no assertion in front of it.

**CI**: a separate `native-build-debug` job with its own baseline file,
because the two configurations compile different code —
`check-native-build-baseline.py` now refuses to compare a debug
measurement against a release baseline instead of ratcheting against the
wrong one.

**A correction to how this repo describes the local replay gate**
(`replay-check-gamedata.md`, new section). The docs implied the retail
replay check simply cannot be run off Windows. It *can* be executed
under wine 11.0 in this repository's own `zerohour-build` image — and
its verdict there is worthless in absolute terms, which is measured, not
inferred: **unmodified `main` mismatches all ten Zero Hour replays at
the identical frames** (9724, 3810, 8310, 3310, 6210, 110, 4810, 4210,
4510, 3611) as this branch, with pinned data hashes and `-jobs 1`
reproducing the earliest. So wine is a *differential* check only, never
an oracle; the authoritative verdict is the `Replay Check GeneralsMD`
job on this PR. Writing this down so the next session does not spend the
run rediscovering it.

**Not measured**: Apple Silicon (all of this is Linux x86-64; the
mechanisms exist on macOS and the existing macOS stack-walk job covers
`backtrace()`, but the debug configuration has not been built on arm64),
and the 13,500-frame replay run under the debug build — that figure came
from the game binary under gdb with the CRC mismatch suppressed, which
is a measurement device, not something to put in CI. No assertion was
disabled, weakened or `#if 0`'d.


Link to Devin session:
https://app.devin.ai/sessions/de290472fc124c10acb607923bbfdc86
Requested by: @willhoff2

+++ GeneralsMD/Code/GameEngine/Source/Common/Recorder.cpp
@@ -92,7 +92,10 @@ static void writeAtOffset(File* file, Int offset, const void* data, Int dataSize
 #if defined(RTS_DEBUG)
 static FILE* openStatsLogFile()
 {
-	unsigned long bufSize = MAX_COMPUTERNAME_LENGTH + 1;
+	// TheSuperHackers @port DWORD, the type GetComputerName() documents, rather than the
+	// `unsigned long` it happens to be on 32-bit Windows. Identical there; off Windows DWORD is
+	// 32 bits wide while `unsigned long` is 64.
+	DWORD bufSize = MAX_COMPUTERNAME_LENGTH + 1;
 	char computerName[MAX_COMPUTERNAME_LENGTH + 1];
 	if (!GetComputerName(computerName, &bufSize))
 	{
